+++ TooManyItems/src/Patches/RehostItemFixes.cs
@@ -1,6 +1,8 @@
-﻿using GameNetcodeStuff;
+﻿using System;
+using GameNetcodeStuff;
 using HarmonyLib;
 using Unity.Netcode;
+using UnityEngine;
 
 namespace TooManyItems.Patches;
 
@@ -8,71 +10,29 @@ namespace TooManyItems.Patches;
 internal class RehostItemFixes
 {
 
-    private static bool _fullyLoaded = false;
-    
-    [HarmonyPrefix]
+    [HarmonyPostfix]
     [HarmonyPatch(typeof(NetworkBehaviour), nameof(NetworkBehaviour.OnNetworkSpawn))]
-    private static void OnNetworkSpawn(NetworkBehaviour __instance)
+    [HarmonyPriority(20)]
+    private static void SpawnPostfix(NetworkBehaviour __instance)
     {
-        if (_fullyLoaded)
+        if (__instance is not GrabbableObject grabbable)
             return;
 
-        var grabbable = __instance as GrabbableObject;
-        if(grabbable == null)
-            return;
-
-        if (grabbable is ClipboardItem || (grabbable is PhysicsProp && grabbable.itemProperties.itemName == "Sticky note"))
+        if (StartOfRound.Instance.localPlayerController != null)
             return;
 
         grabbable.isInElevator = true;
         grabbable.isInShipRoom = true;
+        grabbable.scrapPersistedThroughRounds = true;
         
-        if (grabbable.radarIcon != null)
-            UnityEngine.Object.Destroy(grabbable.radarIcon.gameObject);
-    }
-    
-    [HarmonyPostfix]
-    [HarmonyPatch(typeof(GrabbableObject), nameof(GrabbableObject.Start))]
-    private static void AfterStart(GrabbableObject __instance)
-    {
-        if (__instance.radarIcon != null && __instance.isInShipRoom)
-            UnityEngine.Object.Destroy(__instance.radarIcon.gameObject);
-    }
-    
-    [HarmonyPrefix]
-    [HarmonyPatch(typeof(NetworkBehaviour), nameof(NetworkBehaviour.OnDestroy))]
-    private static void OnDestroy(NetworkBehaviour __instance)
-    {
-        var grabbable = __instance as GrabbableObject;
-        if(grabbable == null)
+        if (grabbable is not LungProp lungProp) 
             return;
         
-        if (grabbable.radarIcon != null)
-            UnityEngine.Object.Destroy(grabbable.radarIcon.gameObject);
-    }
-    
-    [HarmonyPrefix]
-    [HarmonyPatch(typeof(PlayerControllerB), nameof(PlayerControllerB.ConnectClientToPlayerObject))]
-    private static void OnFinishedLoading()
-    {
-        TooManyItems.Log.LogDebug($"{nameof(RehostItemFixes)} player fully loaded! {_fullyLoaded}");
-        _fullyLoaded = true;
-    }
-    
-    [HarmonyPrefix]
-    [HarmonyPatch(typeof(StartOfRound), nameof(StartOfRound.OnLocalDisconnect))]
-    private static void OnDisconnect()
-    {
-        TooManyItems.Log.LogDebug($"{nameof(RehostItemFixes)} reset1! {_fullyLoaded}");
-        _fullyLoaded = false;
-    }
-    
-    [HarmonyPrefix]
-    [HarmonyPatch(typeof(StartOfRound), nameof(StartOfRound.OnDestroy))]
-    private static void OnDestroy()
-    {
-        TooManyItems.Log.LogDebug($"{nameof(RehostItemFixes)} reset2! {_fullyLoaded}");
-        _fullyLoaded = false;
+        lungProp.isLungDocked = false;
+        lungProp.isLungPowered = false;
+        lungProp.isLungDockedInElevator = false;
+        lungProp.GetComponent<AudioSource>()?.Stop();
+
     }
     
 }